Renamed plugin ExampleRssWidget -> RssWidget
