--- plugins/ExampleRssWidget/ExampleRssWidget.php
@@ -1,30 +1,0 @@
-<?php
-/**
- * Piwik - free/libre analytics platform
- *
- * @link http://piwik.org
- * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
- *
- */
-namespace Piwik\Plugins\ExampleRssWidget;
-
-/**
- *
- */
-class ExampleRssWidget extends \Piwik\Plugin
-{
-    /**
-     * @see Piwik\Plugin::registerEvents
-     */
-    public function registerEvents()
-    {
-        return array(
-            'AssetManager.getStylesheetFiles' => 'getStylesheetFiles'
-        );
-    }
-
-    public function getStylesheetFiles(&$stylesheets)
-    {
-        $stylesheets[] = "plugins/ExampleRssWidget/stylesheets/rss.less";
-    }
-}
