+++ plugins/RssWidget/plugin.json
@@ -1,5 +1,5 @@
 {
-    "name": "ExampleRssWidget",
+    "name": "RssWidget",
     "description": "Piwik Platform showcase: how to create a new widget that displays a user submitted RSS feed.",
     "version": "1.0",
     "keywords": ["example", "feed", "widget"],
