+++ tests/PHPUnit/System/expected/test_apiGetReportMetadata__API.getWidgetMetadata_day.xml
@@ -1873,14 +1873,14 @@
 			<id>Example Widgets</id>
 		</category>
 		<subcategory />
-		<module>ExampleRssWidget</module>
+		<module>RssWidget</module>
 		<action>rssChangelog</action>
 		<order>99</order>
 		<parameters>
-			<module>ExampleRssWidget</module>
+			<module>RssWidget</module>
 			<action>rssChangelog</action>
 		</parameters>
-		<uniqueId>widgetExampleRssWidgetrssChangelog</uniqueId>
+		<uniqueId>widgetRssWidgetrssChangelog</uniqueId>
 	</row>
 	<row>
 		<name>Movers and Shakers</name>
@@ -1941,14 +1941,14 @@
 			<id>Example Widgets</id>
 		</category>
 		<subcategory />
-		<module>ExampleRssWidget</module>
+		<module>RssWidget</module>
 		<action>rssPiwik</action>
 		<order>99</order>
 		<parameters>
-			<module>ExampleRssWidget</module>
+			<module>RssWidget</module>
 			<action>rssPiwik</action>
 		</parameters>
-		<uniqueId>widgetExampleRssWidgetrssPiwik</uniqueId>
+		<uniqueId>widgetRssWidgetrssPiwik</uniqueId>
 	</row>
 	<row>
 		<name>Welcome!</name>
